controller and services implement (custom routes)

+++ src/main/java/com/rider/ride/validators/RideValidator.java
@@ -61,7 +61,7 @@ public class RideValidator implements Validator {
         return (input == null);
     }
 
-    private boolean checkDecimal(BigDecimal input) {
+    private boolean checkDecimal(Double input) {
         return (input == null);
     }
 }
